New world class for organization, changed player movement, changes to KeyListener event handling
I created a new world class which is just a Jpanel, this is where the player class is drawn onto. This will be useful later when we have menus and huds and such. It will also be good for world-specific functions like entering new areas and scrolling the screens. Since the keylistener is on the game class, I just used functions with the KeyEvent argument to basically pass down the KeyEvent to the player, from game to world then to player. I have also made it so that holding down a directional key repeats the movement after a set time, unlike before where the time was determined by how fast the keyevent repeats when a key is pressed down.